Issue615 property templates modify values (#619)
Fixed the PropFile and Jndi configuration templates to not write values that actually set a value.  Perviously, an entry might have looked like this:
```
com.bigcorp.MyClass.MY_PROPERTY = [String]
```
Which would set the value of the property to `[String]` if someone using the template didn't edit that line.  Better to leave the value empty so that user's don't have unintended values set.
* Added missing tests
* Clarified some edge cases in sample writing
* rm a few unused methods

Co-authored-by: Eric Everman <eeverman@gmail.com>