Travis will build main (only) on merge
Currently builds are only done on pull requests, however, that resulting in the code coverage report for main falling behind.  The .travis file is now set to only allow branch builds on 'main' and I will turn on branch builds in the Travis settings for this project (on the Travis website).